menambahkan fitur notifikasi ke email dan telegram

+++ app/Http/Controllers/PostController.php
@@ -2,11 +2,16 @@
 
 namespace App\Http\Controllers;
 
+use App\Mail\BlogPosted;
 use App\Models\Post;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\DB;
 
+// semua method di PostController perlu diakses menggunakan login
 class PostController extends Controller
 {
     /**
@@ -15,6 +20,10 @@ class PostController extends Controller
     // NAMA TABLE SEHARUSNYA PLURAL (POSTS)
     public function index() // menampilkan semua
     {
+        if(!Auth::check()) { // sudah login atau belum
+            return redirect("login");
+        }
+
         // active() adalah query scope dari App\Model\Post yaitu scopeActive()
         // withTrashed() adalah scope bawaaan dari SoftDeletes
         $posts = Post::active()->withTrashed()->get();
@@ -30,6 +39,10 @@ class PostController extends Controller
      */
     public function create() // untuk create data lalu submit ke store
     {
+        if(!Auth::check()) { // sudah login atau belum
+            return redirect("login");
+        }
+
         return view("posts.create");
     }
 
@@ -38,12 +51,16 @@ class PostController extends Controller
      */
     public function store(Request $request) // untuk menyimpan data baru
     {
+        if(!Auth::check()) { // sudah login atau belum
+            return redirect("login");
+        }
+
         // membaca request yag dikirimkan client/browser
         $title = $request->input("title");
         $content = $request->input("content");
 
         // Post::insert([ // insert diganti create
-        Post::create([ // sesuaikan dgn nama tabelnya
+        $post = Post::create([ // sesuaikan dgn nama tabelnya
             // field apa saja yg ingin diisi
             "title"=> $title,
             "content"=> $content,
@@ -51,6 +68,13 @@ class PostController extends Controller
             // "created_at" => date("Y-m-d H:i:s"),
             // "updated_at"=> date("Y-m-d H:i:s")
         ]);
+
+        // NOTIFIKASI EMAIL
+        // pengirim adalah email dari yg login di web Blog
+        Mail::to(Auth::user()->email)->send(new BlogPosted($post));
+
+        // NOTIFIKASI TELEGRAM
+        $this->notify_telegram($post); // mengirim objek Post
 
         return redirect("posts");
     }
@@ -60,6 +84,10 @@ class PostController extends Controller
      */
     public function show(string $id) // menampilkan detail / salah satu data saja
     {
+        if(!Auth::check()) { // sudah login atau belum
+            return redirect("login");
+        }
+
         $post = Post::where('id', $id)->first(); // mendapatkan data paling pertama dari query diatasnya (single data, dan harus unik where nya)
         // $post = Post::find($id); // debugging
         // dd($post); // debugging
@@ -81,6 +109,10 @@ class PostController extends Controller
      */
     public function edit(string $id) // menampilkan formnya
     {
+        if(!Auth::check()) { // sudah login atau belum
+            return redirect("login");
+        }
+
         // sama kayak method show bagian ini terus filenya bisa duplikat dari create.blade.php
         $post = Post::where('id', '=', $id)->first(); // mendapatkan data paling pertama dari query diatasnya (single data, dan harus unik where nya)
 
@@ -95,6 +127,10 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id) // melakukan action perubahannya
     {
+        if(!Auth::check()) { // sudah login atau belum
+            return redirect("login");
+        }
+
         $title = $request->input("title");
         $content = $request->input("content");
 
@@ -113,8 +149,28 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
+        if(!Auth::check()) { // sudah login atau belum
+            return redirect("login");
+        }
+
         Post::where("id", $id)->delete();
 
         return redirect("posts");
+    }
+
+    private function notify_telegram($post) {
+        $api_token = "7779606078:AAEj4qCBruLmYmcBBsXJshqRmfdHkU3rSnI"; // BotFather
+        $url = "https://api.telegram.org/bot{$api_token}/sendMessage";
+        $chat_id = -1002314028532; // JsonDumpBot
+
+        $content = "Ada postingan baru nih di blog kamu dengan judul: <strong>\"{$post->title}\"</strong>"; // kutip di dalam kutip
+
+        $data = [
+            "chat_id"       => $chat_id,
+            "text"          => $content,
+            "parse_mode"    => "HTML"
+        ];
+
+        Http::post($url, $data);
     }
 }
